refactor(P2.5 sub-slice B): move CTE-name remap pair into render_plan/cte_rewrite.rs
Phase-2 §5.1 module move, second sub-slice of the cte_rewrite group. Extracted
the tightly-coupled CTE-name remap pair verbatim from plan_builder_utils.rs into
the existing render_plan/cte_rewrite.rs:
remap_cte_names_in_expr (old 1385-1469) + remap_cte_names_in_render_plan (old
1513-1567). Two non-contiguous deletions — they were separated by the unrelated
rewrite_count_to_conditional / quote_qualified_col helpers, which stay put.

`pub(crate) use` re-export for remap_cte_names_in_render_plan (2 internal
callers). remap_cte_names_in_expr has no caller left in plan_builder_utils (it
was only invoked by the two moved fns), so it is not re-exported — it lives on as
a `pub fn` used internally by _in_render_plan. Zero logic edits.

Gates: 1612 lib + 529 integration (incl. corpus_sweep + sql_golden) + 1 ratchet
+ 7 unit, 0 failures; byte-identical (both fn bodies diff-clean vs main, no
golden churn, ratchet net-zero); fmt/clippy clean; warning-free --tests build.
