+++ docs/design/REFACTORING_SAFETY_PLAN.md
@@ -463,7 +463,7 @@ Split along the audited seams, **pure groups first** (lowest risk):
 | `render_plan/pattern_comprehension_sql.rs` | ✅ **moved (P2.2, Jul 2026)** — the pattern-comprehension SQL *string* emitters (31 fns, ~2,600 lines): emits SQL text, a different layer from the rest | cohesive, separable |
 | `render_plan/clause_extractors.rs` | ✅ **moved (P2.3, Jul 2026)** — the pure clause extractors that remained (`extract_having/order_by/limit/skip` + `extract_sorted_properties`, was 1467–1560). NOTE: `extract_filters/from/group_by/distinct` from the original group had already migrated to `filter_builder.rs`/`group_by_builder.rs` via incremental work | the only externally-consumed group; mostly pure |
 | `render_plan/plan_predicates.rs` | ✅ **WITH-detection group moved (P2.4, Jul 2026)** — `has_with_clause_in_tree`/`has_with_clause_in_graph_rel`/`plan_contains_with_clause`. NOTE: the fresh-scan/with-exported alias walkers (`find_fresh_table_scan_aliases_in_plan`/`collect_fresh_scan_aliases`/`with_exported_aliases_in_branch`) are private helpers coupled to P2.5's cte_rewrite fns, so they ride with P2.5, not here | pure read-only walkers — natural home for the §4 `walk()` rewrites |
-| `render_plan/cte_rewrite.rs` | ◐ **in progress (P2.5, sub-sliced)** — CTE-ref extraction + CTE/alias rewriting. Sub-slice A moved the expression-rewriting cluster (`rewrite_operator_application_for_cte`/`_join`, `rewrite_render_expr_for_cte_simple`/`_operand`). Remaining: `remap_cte_names_*`, `rewrite_logical_expr_cte_refs`, `update_graph_joins_cte_refs`, deferred P2.4 alias walkers, D2 dedup | pure-ish (RenderPlan/RenderExpr + maps) |
+| `render_plan/cte_rewrite.rs` | ◐ **in progress (P2.5, sub-sliced)** — CTE-ref extraction + CTE/alias rewriting. Sub-slice A: expression-rewriting cluster (`rewrite_operator_application_for_cte`/`_join`, `rewrite_render_expr_for_cte_simple`/`_operand`). Sub-slice B: CTE-name remap pair (`remap_cte_names_in_expr`/`_in_render_plan`). Remaining: `collect_with_cte_table_aliases`, `rewrite_join_conditions_for_cte_aliases`, `rewrite_logical_expr_cte_refs`, `update_graph_joins_cte_refs`, deferred P2.4 alias walkers, D2 dedup | pure-ish (RenderPlan/RenderExpr + maps) |
 | `render_plan/with_to_cte/` (dir) | 6999–15491: the two giant builders + their orbit | **entangled core** — moved in Phase 2, decomposed in Phase 4 |
 
 Rules for the move slices: `pub(crate)` re-exports from the old path during
@@ -1006,9 +1006,12 @@ byte-identical corpus + goldens, ratchet net-zero) ·
 cluster (`rewrite_operator_application_for_cte`/`_for_cte_join` +
 `rewrite_render_expr_for_cte_simple`/`_operand`) → `render_plan/cte_rewrite.rs`,
 `pub(crate)` re-exports, byte-identical (one `fn`→`pub(crate) fn` widening),
-ratchet net-zero. Remaining P2.5: `remap_cte_names_*`,
-`rewrite_logical_expr_cte_refs`, `update_graph_joins_cte_refs`, the deferred
-P2.4 alias walkers, D2 dedup · ☐ P2.6 with_to_cte move · ☐ P2.7 D1 ·
+ratchet net-zero. Sub-slice B done: the CTE-name remap pair
+(`remap_cte_names_in_expr`/`_in_render_plan`) → same module, byte-identical,
+ratchet net-zero. Remaining P2.5: `collect_with_cte_table_aliases`,
+`rewrite_join_conditions_for_cte_aliases`, `rewrite_logical_expr_cte_refs`,
+`update_graph_joins_cte_refs`, the deferred P2.4 alias walkers, D2 dedup ·
+☐ P2.6 with_to_cte move · ☐ P2.7 D1 ·
 ☐ P2.8 D6 · ☐ P2.9 D8 · ☐ P2.10 import hygiene + remaining dead_code
 
 Phase 3: ◐ Slices 1–2 merged 2026-07-13: GraphNode/ViewScan + query_planner
